and only fragile scaffolding will autobreak

+++ src/main/java/com/lothrazar/cyclicmagic/block/BlockScaffoldingReplace.java
@@ -1,9 +1,9 @@
 package com.lothrazar.cyclicmagic.block;
-import javax.annotation.Nullable;
 import com.lothrazar.cyclicmagic.IHasRecipe;
 import net.minecraft.block.Block;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.entity.player.EntityPlayer;
+import net.minecraft.init.Blocks;
 import net.minecraft.init.Items;
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.EnumFacing;
@@ -14,16 +14,18 @@ import net.minecraftforge.fml.common.registry.GameRegistry;
 
 public class BlockScaffoldingReplace extends BlockScaffolding implements IHasRecipe {
   public BlockScaffoldingReplace() {
-    super();
+    super(false);
     this.dropBlock = true;
   }
-  public boolean onBlockActivated(World worldIn, BlockPos pos, IBlockState state, EntityPlayer playerIn, EnumHand hand, @Nullable ItemStack heldItem, EnumFacing side, float hitX, float hitY, float hitZ) {
-    if (heldItem == null) { return false; }
+  @Override
+  public boolean onBlockActivated(World worldIn, BlockPos pos, IBlockState state, EntityPlayer playerIn, EnumHand hand, EnumFacing side, float hitX, float hitY, float hitZ) {
+    ItemStack heldItem = playerIn.getHeldItem(hand);
+    if (heldItem.isEmpty()) { return false; }
     Block b = Block.getBlockFromItem(heldItem.getItem());
-    if (b != null && !(b instanceof BlockScaffolding)) {
+    if (b != null && b != Blocks.AIR && !(b instanceof BlockScaffolding)) {
       worldIn.destroyBlock(pos, dropBlock);
-      //     worldIn.setBlockState(pos, Block.getBlockFromItem(heldItem.getItem()) .getDefaultState());
-      heldItem.onItemUse(playerIn, worldIn, pos, hand, side, hitX, hitY, hitZ); //.useItemRightClick(worldIn, playerIn, hand);
+      //     worldIn.setBlockState(pos, Block.getBlockFromItem(heldItem.getItem()) .getDefaultState()); //.useItemRightClick(worldIn, playerIn, hand);ee
+      heldItem.onItemUse(playerIn, worldIn, pos, hand, side, hitX, hitY, hitZ);
       return true;//to cancel event chains
     }
     return false;
